1. Fix up changes needed on Master

+++ bin/tsx_cmd_install.sh
@@ -42,22 +42,25 @@ export ROOT_URL='http://127.0.0.1'
 
 if [ "$(uname)" == "Darwin" ]; then
     echo Mac in ${install_dir}
+    if [ "${1}" != "update-only" ]; then
 
-    echo " *******************************"
-    echo "Mongodb - Download and extract"
-    echo " *******************************"
-    curl https://fastdl.mongodb.org/osx/mongodb-osx-ssl-x86_64-4.0.0.tgz -o mongodb-osx-ssl-x86_64-4.0.0.tgz
-    tar -xf mongodb-osx-ssl-x86_64-4.0.0.tgz
-    rm ${install_dir}/mongodb-osx-ssl-x86_64-4.0.0.tgz
+      # install mongodb
+      echo " *******************************"
+      echo "Mongodb - Download and extract"
+      echo " *******************************"
+      curl https://fastdl.mongodb.org/osx/mongodb-osx-ssl-x86_64-4.0.0.tgz -o mongodb-osx-ssl-x86_64-4.0.0.tgz
+      tar -xf mongodb-osx-ssl-x86_64-4.0.0.tgz
+      rm ${install_dir}/mongodb-osx-ssl-x86_64-4.0.0.tgz
 
-    # install nodejs
-    echo " *******************************"
-    echo "nodejs - Download and extract"
-    echo " *******************************"
-    curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-darwin-x64.tar.gz -o node-v8.11.3-darwin-x64.tar.gz
-    tar -xf node-v8.11.3-darwin-x64.tar.gz
-    rm ${install_dir}/node-v8.11.3-darwin-x64.tar.gz
-
+      # install nodejs
+      echo " *******************************"
+      echo "nodejs - Download and extract"
+      echo " *******************************"
+      curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-darwin-x64.tar.gz -o node-v8.11.3-darwin-x64.tar.gz
+      tar -xf node-v8.11.3-darwin-x64.tar.gz
+      rm ${install_dir}/node-v8.11.3-darwin-x64.tar.gz
+    fi
+    # https://drive.google.com/drive/folders/1yUPU6A0gbBv5UnuSp308lvctY4d6aUtw?usp=sharing
     export PATH=${install_dir}/mongodb-osx-x86_64-4.0.0/bin:${install_dir}/node-v8.11.3-darwin-x64/bin:$PATH
 
 elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
@@ -67,53 +70,51 @@ elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
     # install nodejs
     if [ "$(uname -p)" == "aarch64" ]; then
 
-
-      echo " *******************************"
-      echo "Mongodb - apt-get install"
-      echo " *******************************"
-      # source: https://andyfelong.com/2018/02/update-mongodb-3-6-on-odroid-c2-with-ubuntu-16-04-3-arm64/
-      sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv 2930ADAE8CAF5059EE73BB4B58712A2291FA4AD5
-      echo "deb [ arch=amd64,arm64,ppc64el,s390x ] http://repo.mongodb.com/apt/ubuntu xenial/mongodb-enterprise/3.6 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-enterprise.list
-      sudo apt update
-      sudo apt install g++ build-essential -y
-      sudo apt install mongodb-enterprise -y
-
-      echo " *******************************"
-      echo "nodejs - Download and extract"
-      echo " *******************************"
-      # armv7
-      curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-linux-armv7l.tar.xz -o node-v8.11.3-linux-armv7l.tar.xz
-      tar -xf node-v8.11.3-linux-armv7l.tar.xz
-      rm ${install_dir}/node-v8.11.3-linux-armv7l.tar.xz
-
+      if [ "${1}" != "update-only" ]; then
+        # install mongodb
+        echo " *******************************"
+        echo "Mongodb - apt-get install"
+        echo " *******************************"
+        # source: https://andyfelong.com/2018/02/update-mongodb-3-6-on-odroid-c2-with-ubuntu-16-04-3-arm64/
+        sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv 2930ADAE8CAF5059EE73BB4B58712A2291FA4AD5
+        echo "deb [ arch=amd64,arm64,ppc64el,s390x ] http://repo.mongodb.com/apt/ubuntu xenial/mongodb-enterprise/3.6 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-enterprise.list
+        sudo apt update
+        sudo apt install g++ build-essential -y
+        sudo apt install mongodb-enterprise -y
+        
+        echo " *******************************"
+        echo "nodejs - Download and extract"
+        echo " *******************************"
+        # armv7
+        curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-linux-armv7l.tar.xz -o node-v8.11.3-linux-armv7l.tar.xz
+        tar -xf node-v8.11.3-linux-armv7l.tar.xz
+        rm ${install_dir}/node-v8.11.3-linux-armv7l.tar.xz
+      fi
       export PATH=${install_dir}/node-v8.11.3-linux-armv7l/bin:$PATH
       # armv8
       #curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-linux-arm64.tar.xz -o node-v8.11.3-linux-arm64.tar.xz
       #tar -xf node-v8.11.3-linux-arm64.tar.xz
       #rm ${install_dir}/node-v8.11.3-linux-arm64.tar.xz
     elif [ "$(uname -p)" == "armv7l" ]; then
+      # install mongodb
+      if [ "${1}" != "update-only" ]; then
+        echo " *******************************"
+        echo "Mongodb - BUILD - NEED TO BE BUILT FOR ARMv7"
+        echo "run the following (it can take a while on ODroid-XU4)"
+        echo " ./mongod_arm_build.sh"
+        echo " *******************************"
 
-      echo " *******************************"
-      echo "         WARNING WARNING"
-      echo " *******************************"
-      echo "Mongodb - BUILD - NEED TO BE BUILT FOR ARMv7"
-      echo "run the following (it can take a while on ODroid-XU4)"
-      echo " ./bin/mongod_armv7_build.sh"
-      echo " *******************************"
-      echo "         WARNING WARNING"
-      echo " *******************************"
-      echo ""
-      echo " *******************************"
-      echo "nodejs - Download and extract"
-      #curl https://nodejs.org/dist/latest-v6.x/node-v6.14.4-linux-armv7l.tar.gz -o node-v6.14.4-linux-armv7l.tar.gz
-      echo " *******************************"
-      curl https://nodejs.org/dist/v6.14.4/node-v6.14.4-linux-armv7l.tar.gz  -o node-v6.14.4-linux-armv7l.tar.gz
-      tar -xf node-v6.14.4-linux-armv7l.tar.gz
-      rm ${install_dir}/node-v6.14.4-linux-armv7l.tar.gz
-      # FIX for odroid - 2018-11-16 no longer needed
-      #rm -rf ~/.node-gyp
-      #npm install -g node-gyp
-
+        echo " *******************************"
+        echo "nodejs - Download and extract"
+        echo " *******************************"
+        # armv7
+        curl https://nodejs.org/dist/latest-v6.x/node-v6.14.4-linux-armv7l.tar.gz -o node-v6.14.4-linux-armv7l.tar.gz
+        tar -xf node-v6.14.4-linux-armv7l.tar.gz
+        rm ${install_dir}/node-v6.14.4-linux-armv7l.tar.gz
+        # FIX for odroid - 2018-11-16 no longer needed
+        #rm -rf ~/.node-gyp
+        #npm install -g node-gyp
+      fi
       export PATH=${install_dir}/node-v6.14.4-linux-armv7l/bin:$PATH
     else
       echo NO NODEJS supported
@@ -126,19 +127,22 @@ else
 fi
 
 cd ${install_dir}
-
-echo " *******************************"
-echo " TSX_CMD - Extract" ${1}
-echo " *******************************"
-# https://drive.google.com/drive/folders/1yUPU6A0gbBv5UnuSp308lvctY4d6aUtw?usp=sharing
-tar -xf ${1}
+if [ "${1}" != "update-only" ]; then
+  # install tsx__cmd - assumes already downloaded
+  echo " *******************************"
+  echo " TSX_CMD - Extract" ${1}
+  echo " *******************************"
+  # https://drive.google.com/drive/folders/1yUPU6A0gbBv5UnuSp308lvctY4d6aUtw?usp=sharing
+    tar -xf ${1}
+fi
 
 cd ${install_dir}/bundle/programs/server
-echo " *******************************"
-echo " TSX_CMD - update npm installs"
-echo " *******************************"
-npm install amdefine ansi-styles chalk escape-string-regexp has-ansi promise source-map strip-ansi type-of ansi-regex asap eachline meteor-promise semver source-map-support supports-color underscore
-
+if [ "${1}" != "update-only" ]; then
+  echo " *******************************"
+  echo " TSX_CMD - update npm installs"
+  echo " *******************************"
+  npm install amdefine ansi-styles chalk escape-string-regexp has-ansi promise source-map strip-ansi type-of ansi-regex asap eachline meteor-promise semver source-map-support supports-color underscore
+fi
 echo " *******************************"
 echo " TSX_CMD - fix for fibers deploy"
 echo " *******************************"
@@ -147,32 +151,5 @@ npm install fibers
 cd ${install_dir}
 
 echo " *******************************"
-echo " TSX_CMD - Create settings file"
-echo " *******************************"
-touch ${install_dir}/settings.json
-cat > ${install_dir}/settings.json <<- EOF
-{
-  "enable_log": "yes",
-  "enable_debug": "no",
-  "enable_info": "no",
-  "enable_warn": "yes",
-  "removed-to-enable: log_file_location": "/media/odroid/PENSIVE2/tsx_cmd_logs"
-}
-EOF
-
-echo " *******************************"
 echo " TSX_CMD - installed"
 echo " *******************************"
-
-if [ "$(uname -p)" == "armv7l" ]; then
-  echo " *******************************"
-  echo "         WARNING WARNING"
-  echo " *******************************"
-  echo "Mongodb - BUILD - NEED TO BE BUILT FOR ARMv7"
-  echo "run the following (it can take a while on ODroid-XU4)"
-  echo " ./bin/mongod_armv7_build.sh"
-  echo " *******************************"
-  echo "         WARNING WARNING"
-  echo " *******************************"
-  echo ""
-fi
